YKI(Frontend): Refactor Move listings under listing folder
