+++ frontend/packages/yki/clerk/src/components/clerkQuarantine/listing/ClerkQuarantineModal.tsx
